live-normalizer/requirements.txt
pyzmq>=25.0
confluent-kafka>=2.3
requests>=2.31
